# gateway/docs/swagger.yaml
definitions:
...
  genproto.AddPaymentReq:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      payment_method:
        type: string
      status:
        type: string
      transaction_id:
        type: string
    type: object
...
  genproto.Booking:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/genproto.GeoPoint'
      provider_id:
        $ref: '#/definitions/genproto.Provider'
      schudule_time:
        type: string
      service_id:
        $ref: '#/definitions/genproto.Services'
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  genproto.GetProviderResp:
    properties:
      provider:
        $ref: '#/definitions/genproto.Provider'
    type: object
...
  genproto.ListAllPaymentsResp:
    properties:
      payments:
        items:
          $ref: '#/definitions/genproto.Payment'
        type: array
    type: object
  genproto.ListAllProvidersResp:
    properties:
      providers:
        items:
          $ref: '#/definitions/genproto.Provider'
        type: array
    type: object
...
  genproto.Payment:
    properties:
      amount:
        type: number
      booking_id:
        $ref: '#/definitions/genproto.Booking'
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      payment_method:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
    type: object
  genproto.Provider:
    properties:
      availability:
        type: string
      average_rating:
        type: number
      company_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/genproto.GeoPoint'
      services:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  genproto.RegisterProviderReq:
    properties:
      availability:
        type: string
      average_rating:
        type: number
      company_name:
        type: string
      description:
        type: string
      location:
        $ref: '#/definitions/genproto.GeoPoint'
      services:
        items:
          type: string
        type: array
      user_id:
        type: string
    type: object
...
  genproto.SearchProvidersResp:
    properties:
      providers:
        items:
          $ref: '#/definitions/genproto.Provider'
        type: array
    type: object
...
  genproto.Services:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  genproto.UpdateProviderReq:
    properties:
      id:
        type: string
      provider:
        $ref: '#/definitions/genproto.Provider'
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/v1/payment:
    get:
      consumes:
      - application/json
      description: This API retrieves all Payments with optional filters
      parameters:
      - description: Booking ID
        in: query
        name: booking_id
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ListAllPaymentsResp'
        "400":
          description: 'error": "error message'
          schema:
            type: string
      summary: GET all Payments
      tags:
      - carwash/Payment
  /api/v1/payment/{id}:
    get:
      consumes:
      - application/json
      description: This API retrieves a Payment by ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Payment'
        "400":
          description: 'error": "error message'
          schema:
            type: string
      summary: GET Payment
      tags:
      - carwash/Payment
  /api/v1/payment/add:
    post:
      consumes:
      - application/json
      description: This API creates a Payment
      parameters:
      - description: Payment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.AddPaymentReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "created successfully'
          schema:
            type: string
        "400":
          description: 'error": "error message'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: CREATE Payment
      tags:
      - carwash/Payment
  /api/v1/provider:
    get:
      consumes:
      - application/json
      description: This API Lists a new provider
      parameters:
      - description: Company Name
        in: query
        name: company_name
        type: string
      - description: Description
        in: query
        name: description
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ListAllProvidersResp'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all provider
      tags:
      - carwash/provider
  /api/v1/provider/{id}:
    delete:
      consumes:
      - application/json
      description: This API deleted a new provider
      parameters:
      - description: provider Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: provider deleted successfully
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete provider
      tags:
      - carwash/provider
    get:
      consumes:
      - application/json
      description: This API Gets a  provider
      parameters:
      - description: provider Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.GetProviderResp'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get provider
      tags:
      - carwash/provider
    put:
      consumes:
      - application/json
      description: This API Put s a new provider
      parameters:
      - description: Put  provider Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateProviderReq'
      produces:
      - application/json
      responses:
        "200":
          description: provider updated successfully
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Put  provider
      tags:
      - carwash/provider
  /api/v1/provider/add:
    post:
      consumes:
      - application/json
      description: This API creates a provider
      parameters:
      - description: Provider
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.RegisterProviderReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "created successfully'
          schema:
            type: string
        "400":
          description: 'error": "error message'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: CREATE provider
      tags:
      - carwash/provider
  /api/v1/provider/search:
    get:
      consumes:
      - application/json
      description: This API gets a  provider
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: description
        in: path
        name: description
        required: true
        type: string
      - description: user_id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.SearchProvidersResp'
        "400":
          description: error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get provider
      tags:
      - carwash/provider
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"